requirements.txt
yt-dlp>=2024.1.1
python-dotenv>=1.0.0

